Switch numbers over to f64
This fixes a few tests that were failing due to not being able to parse
decimal literals, and also fixes the nan_equality test.

I had to add a method to the cursor that I don't particularly like, but
I'll keep thinking about how to make that nice.

+++ src/bytecode/chunk.rs
@@ -11,7 +11,7 @@ use super::string_interning::InternedString;
 // We do allow string interning though, just for convenience.
 #[derive(Clone)]
 pub enum ChunkConstant {
-    Number(u32),
+    Number(f64),
     String(InternedString),
     FnTemplate {
         name: InternedString,
